+++ src/provider/Diagnostic/DA/CDiagFnLib/C501Class.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import type { TParamMapOut } from '../../../../AhkSymbol/CAhkFunc';
 import { DiagsDA, EDiagCodeDA } from '../../../../diag';
 import { CDiagFn } from '../../tools/CDiagFn';
 
@@ -14,5 +15,31 @@ export class C501Class extends CDiagFn {
             tags: [vscode.DiagnosticTag.Unnecessary],
             message: DiagsDA[EDiagCodeDA.code501].msg,
         });
+    }
+}
+
+export function NeverUsedParam(
+    paramMap: TParamMapOut,
+    code501List: C501Class[],
+    displayErrList: readonly boolean[],
+): void {
+    for (const v of paramMap.values()) {
+        const {
+            isByRef,
+            refRangeList,
+            keyRawName,
+            defRangeList,
+        } = v;
+
+        if (
+            isByRef
+            || refRangeList.length > 0
+            || keyRawName.startsWith('_')
+            || !displayErrList[defRangeList[0].start.line]
+        ) {
+            continue;
+        }
+
+        code501List.push(new C501Class(defRangeList));
     }
 }