+++ src/tools/DeepAnalysis/FnVar/def/OutputVarCommandPseudoArray.ts
@@ -1,5 +1,0 @@
-/**
- * <https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/issues/11>
- */
-
-const z = 0;