+++ src/tools/DeepAnalysis/FnVar/def/wrapFnValDef.ts
@@ -1,5 +1,5 @@
 import type * as vscode from 'vscode';
-import type { TValMapIn, TValMetaIn } from '../../../../AhkSymbol/CAhkFunc';
+import type { TAssociated, TValMapIn, TValMetaIn } from '../../../../AhkSymbol/CAhkFunc';
 import { ToUpCase } from '../../../str/ToUpCase';
 import type { EFnMode } from '../EFnMode';
 import { newC502 } from './c502';
@@ -10,6 +10,7 @@ type TGetValue = {
     defRange: vscode.Range,
     lineComment: string,
     fnMode: EFnMode,
+    Associated: TAssociated | null,
 };
 
 export function wrapFnValDef({
@@ -18,6 +19,7 @@ export function wrapFnValDef({
     defRange,
     lineComment,
     fnMode,
+    Associated,
 }: TGetValue): TValMetaIn {
     const oldVal: TValMetaIn | undefined = valMap.get(ToUpCase(RawNameNew));
     if (oldVal !== undefined) {
@@ -26,8 +28,12 @@ export function wrapFnValDef({
         if (lineComment.startsWith(';')) {
             oldVal.commentList.push(lineComment.slice(1));
         }
+        if (Associated !== null) {
+            oldVal.AssociatedList.push(Associated);
+        }
         return oldVal;
     }
+
     return {
         keyRawName: RawNameNew,
         defRangeList: [defRange],
@@ -40,5 +46,8 @@ export function wrapFnValDef({
         ],
         fnMode,
         jsDocStyle: '', // default with ''
+        AssociatedList: Associated === null
+            ? []
+            : [Associated],
     };
 }